@@ -1,12 +1,12 @@
 name: vertical_weight_slider_example
 description: Demonstrates how to use the vertical_weight_slider plugin.
-
+version: 2.0.0
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
