@@ -1,10 +1,10 @@
 name: vertical_weight_slider
 description: This package provides easy-to-use and scrollable vertical weight slider. Compatible with Android & iOS & Web.
-version: 1.0.0
+version: 2.0.0
 homepage: https://github.com/mincheol-shin/vertical_weight_slider
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
