@@ -33,11 +33,11 @@ class _MyAppState extends State<MyApp> {
             ),
             VerticalWeightSlider(
               maximumWeight: 200,
-              initialWeight: 0,
+              initialWeight: 50,
               gradationColor: [
                 Colors.grey[500],
-                Colors.grey[400],
-                Colors.grey[300]
+                Colors.grey[300],
+                Colors.grey[200]
               ],
               onChanged: (value) {
                 setState(() {
